+++ kgcn/src/models/learners_test.py
@@ -18,8 +18,8 @@ class TestSupervisedLoss(unittest.TestCase):
             with self.subTest(labels=str(label_value)):
                 num_samples = 30
                 raw_labels = [label_value for _ in range(num_samples)]
-                labels = tf.convert_to_tensor(raw_labels, dtype=tf.float64)
-                predictions = tf.convert_to_tensor(raw_labels, dtype=tf.float64)
+                labels = tf.convert_to_tensor(raw_labels, dtype=tf.float32)
+                predictions = tf.convert_to_tensor(raw_labels, dtype=tf.float32)
 
                 loss = base.supervised_loss(predictions, labels, 0.0, True)
                 print(loss)
